# docs/migration/contract/openapi-legacy-baseline.yaml
openapi: 3.1.0
info:
  title: Gia Pha Legacy Backend HTTP Contract (Frozen Baseline)
  version: 1.0.0-legacy-baseline
  description: |
    Frozen contract of every Next.js App Router `/api/**` handler prior to the
    Spring Boot migration (Task 1 of spring-boot-backend-migration).

    Conventions frozen from the legacy implementation:
    * **Error envelope** — every error response is `{"ok": false, "error": {"code", "message", "details?"}}`.
    * **Mixed success envelopes** — most successful responses return the raw
      resource JSON (no envelope). The exceptions that wrap in
      `{"ok": true, "data": ...}` are: `POST /api/auth/register`,
      `PATCH /api/trees/{treeId}/memberships/{userId}`,
      `POST /api/media/upload`, and share-link management responses.
    * **Timestamps** — ISO-8601 UTC strings with milliseconds (`.toISOString()` form).
    * **Authentication** — NextAuth JWT session cookie
      (`next-auth.session-token` / `__Secure-next-auth.session-token`).
      Unauthenticated requests receive `401 UNAUTHORIZED`.
    * **RBAC** — tree-scoped roles ADMIN/EDITOR/VIEWER, owner is immutable
      effective ADMIN. Missing tree → `404 TREE_NOT_FOUND`, missing
      permission → `403 FORBIDDEN`.
    * **Tree discovery** — detail routes for members/relationships/events/media/
      albums accept optional `?treeId=`; without it the legacy handler scans
      every tree to locate the resource. Compatibility_API must preserve this.
...
servers:
  - url: /
...
paths:
...
  /api/auth/register:
    post:
      tags: [auth]
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterInput' }
      responses:
        '201':
          description: Account created. Vietnamese message text is part of the contract.
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { const: true }
                  data:
                    type: object
                    required: [user, emailVerificationRequired, message]
                    properties:
                      user:
                        type: object
                        required: [id, email, name, createdAt]
                        properties:
                          id: { type: string }
                          email: { type: string }
                          name: { type: string }
                          createdAt: { type: string, format: date-time }
                      emailVerificationRequired: { type: boolean }
                      message:
                        type: string
                        description: >-
                          "Tài khoản đã được tạo. Vui lòng kiểm tra email để xác nhận tài khoản."
                          or "Tài khoản đã được tạo. Bạn có thể đăng nhập ngay."
        '400': { $ref: '#/components/responses/ValidationError' }
        '409':
          description: EMAIL_ALREADY_EXISTS
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        '503':
          description: EMAIL_DELIVERY_FAILED (any non-conflict AuthServiceError)
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/auth/verify-email:
    get:
      tags: [auth]
      operationId: verifyEmail
      security: []
      parameters:
        - { name: token, in: query, required: false, schema: { type: string } }
      responses:
        '307':
          description: |
            Always redirects to `/vi/login`. Success appends `?verified=1`;
            failure appends `?error=INVALID_TOKEN` (or the AuthServiceError code).
          headers:
            Location: { schema: { type: string } }
...
  /api/events/{eventId}:
    parameters:
      - { name: eventId, in: path, required: true, schema: { type: string } }
      - { name: treeId, in: query, required: false, schema: { type: string } }
    get:
      tags: [events]
      operationId: getEvent
      responses:
        '200': { description: 'Event plus members[] and media[].', content: { application/json: { schema: { $ref: '#/components/schemas/EventDetails' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [events]
      operationId: updateEvent
      description: Merged patch re-validated as a full event (cross-field rules apply); no-op returns current.
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateEventInput' } } } }
      responses:
        '200': { description: Raw updated event., content: { application/json: { schema: { $ref: '#/components/schemas/Event' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [events]
      operationId: deleteEvent
      responses:
        '204': { description: Deleted. Empty body. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------------------------------------------------------- media
...
  /api/media/{mediaId}:
    parameters:
      - { name: mediaId, in: path, required: true, schema: { type: string } }
      - { name: treeId, in: query, required: false, schema: { type: string } }
    get:
      tags: [media]
      operationId: getMedia
      responses:
        '200': { description: Raw metadata., content: { application/json: { schema: { $ref: '#/components/schemas/MediaMetadata' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [media]
      operationId: deleteMedia
      description: Metadata removed first; blob cleanup failures logged, not surfaced.
      responses:
        '204': { description: Deleted. Empty body. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/albums/{albumId}:
    parameters:
      - { name: albumId, in: path, required: true, schema: { type: string } }
      - { name: treeId, in: query, required: false, schema: { type: string } }
    patch:
      tags: [albums]
      operationId: updateAlbum
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateAlbumInput' } } } }
      responses:
        '200': { description: Raw updated album., content: { application/json: { schema: { $ref: '#/components/schemas/Album' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [albums]
      operationId: deleteAlbum
      description: Album removed; linked media keep existing but albumId is cleared.
      responses:
        '204': { description: Deleted. Empty body. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------------------------------------------------------- search
  /api/search:
    get:
      tags: [search]
      operationId: searchMembers
      description: |
        Requires READ on treeId. Accent-insensitive Vietnamese matching
        (NFD strip + đ→d). Modes: search (default), autocomplete, filter.
        Scoring frozen: fieldWeight fullName=30/nickname=20/other=10 plus
        matchWeight exact=100/prefix=70/word-start=50/substring=30; ties by
        source order. Autocomplete ranks 140/135/120/115/100/95, min 2 chars.
        Conflicting alive-status filters → 400 INVALID_FILTER.
      parameters:
        - { name: treeId, in: query, required: true, schema: { type: string } }
        - { name: mode, in: query, required: false, schema: { type: string, enum: [search, autocomplete, filter], default: search } }
        - { name: q, in: query, required: false, schema: { type: string } }
        - { name: fields, in: query, required: false, schema: { type: string }, description: 'CSV of fullName,nickname,occupation,placeOfBirth' }
        - { name: gender, in: query, required: false, schema: { type: string }, description: CSV }
        - { name: generation, in: query, required: false, schema: { type: string }, description: CSV }
        - { name: birthYearFrom, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 9999 } }
        - { name: birthYearTo, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 9999 } }
        - { name: isAlive, in: query, required: false, schema: { type: boolean } }
        - { name: status, in: query, required: false, schema: { type: string, enum: [ALIVE, DECEASED] } }
        - { name: location, in: query, required: false, schema: { type: string } }
        - { name: offset, in: query, required: false, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 0, maximum: 1000 } }
      responses:
        '200':
          description: SearchResult[] | AutocompleteItem[] | Member[] depending on mode.
          content: { application/json: { schema: { type: array, items: {} } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------------------------------------------------------- import / export
...
components:
  securitySchemes:
    nextAuthSession:
      type: apiKey
      in: cookie
      name: next-auth.session-token
...
  responses:
    Unauthorized:
      description: Missing/invalid session (code UNAUTHORIZED).
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    Forbidden:
      description: Authenticated but lacking the tree permission (code FORBIDDEN).
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    NotFound:
      description: Resource or tree not found (code NOT_FOUND / TREE_NOT_FOUND).
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    ValidationError:
      description: Invalid input (code VALIDATION_ERROR / INVALID_INPUT / INVALID_FILTER; details carries Zod flatten()).
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    InternalError:
      description: Unexpected failure (code INTERNAL_ERROR); internals never leak.
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
  schemas:
    ErrorEnvelope:
      type: object
      required: [ok, error]
      properties:
        ok: { const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
...
    Gender: { type: string, enum: [MALE, FEMALE, OTHER] }
...
    EventType: { type: string, enum: [BIRTHDAY, DEATH_ANNIVERSARY, ANNIVERSARY, WEDDING, FUNERAL, REUNION, CUSTOM] }
...
    Member:
      type: object
      required: [id, treeId, firstName, lastName, fullName, gender, isAlive, createdAt, updatedAt]
      properties:
        id: { type: string }
        treeId: { type: string }
        firstName: { type: string, minLength: 1, maxLength: 100 }
        lastName: { type: string, minLength: 1, maxLength: 100 }
        fullName: { type: string, minLength: 1, maxLength: 200 }
        nickname: { type: string, maxLength: 100 }
        gender: { $ref: '#/components/schemas/Gender' }
        dateOfBirth: { type: string, description: ISO date or date-time }
        dateOfDeath: { type: string }
        isAlive: { type: boolean }
        placeOfBirth: { type: string, maxLength: 500 }
        currentAddress: { type: string, maxLength: 500 }
        occupation: { type: string, maxLength: 200 }
        education: { type: string, maxLength: 200 }
        phone: { type: string, maxLength: 20 }
        email: { type: string, maxLength: 254 }
        biography: { type: string, maxLength: 5000 }
        avatarUrl: { type: string, deprecated: true }
        avatarMediaId: { type: string, description: Must reference same-tree image media }
        generation: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Event:
      type: object
      required: [id, treeId, title, type, eventDate, memberIds, mediaIds, createdAt, updatedAt]
      properties:
        id: { type: string }
        treeId: { type: string }
        title: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 5000 }
        type: { $ref: '#/components/schemas/EventType' }
        customType: { type: string, maxLength: 100 }
        eventDate: { type: string }
        location: { type: string, maxLength: 500 }
        memberIds: { type: array, items: { type: string }, maxItems: 1000 }
        mediaIds: { type: array, items: { type: string }, maxItems: 1000 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    UpdateEventInput: { type: object, description: Partial; merged value re-validated as full event. }
    EventDetails:
      allOf:
        - { $ref: '#/components/schemas/Event' }
        - type: object
          properties:
            members: { type: array, items: { $ref: '#/components/schemas/Member' } }
            media: { type: array, items: { $ref: '#/components/schemas/MediaMetadata' } }
    MediaMetadata:
      type: object
      required: [id, treeId, memberIds, eventIds, filename, originalName, mimeType, fileSize, blobUrl, uploadedAt]
      properties:
        id: { type: string }
        treeId: { type: string }
        memberId: { type: string, deprecated: true }
        eventId: { type: string, deprecated: true }
        memberIds: { type: array, items: { type: string } }
        eventIds: { type: array, items: { type: string } }
        albumId: { type: string }
        filename: { type: string }
        originalName: { type: string, maxLength: 255 }
        mimeType: { type: string, enum: [image/jpeg, image/png, image/webp, application/pdf] }
        fileSize: { type: integer, maximum: 10485760 }
        blobUrl: { type: string, description: Private storage URL — never usable directly by clients }
        thumbnailUrl: { type: string }
        contentUrl: { type: string, description: '/api/media/{id}/content?treeId=...' }
        thumbnailContentUrl: { type: string }
        caption: { type: string, maxLength: 2000 }
        takenAt: { type: string, format: date-time }
        uploadedAt: { type: string, format: date-time }
    Album:
      type: object
      required: [id, treeId, name, createdAt, updatedAt]
      properties:
        id: { type: string }
        treeId: { type: string }
        name: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 2000 }
        coverMediaId: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    UpdateAlbumInput: { type: object }
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string, minLength: 2, maxLength: 100 }
        email: { type: string, format: email, maxLength: 254, description: Lowercased before storage }
        password: { type: string, minLength: 12, maxLength: 72, description: Requires upper, lower, digit and special character }
...
security:
  - nextAuthSession: []